Closet/requirements.txt
asgiref==3.2.10
bcrypt==3.1.7
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
defusedxml==0.6.0
Django==3.0.6
django-allauth==0.42.0
django-rest-auth==0.9.5
django-smtp-ssl==1.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
idna==2.10
mysqlclient==1.4.6
oauthlib==3.1.0
pycparser==2.20
PyJWT==1.7.1
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9
